python_code/requirements.txt
datetime
elasticsearch
kafka-python
